+++ packages/ui/src/components/button/index.ts
@@ -1,2 +1,2 @@
-export type { ButtonProps } from './button';
+export type { ButtonProps, ButtonSize, ButtonVariant } from './button';
 export { Button, buttonVariants } from './button';